requirements.txt
fastapi>=0.78.0
uvicorn[standard]>=0.17.0
python-dotenv>=0.19.0
PyPDF2>=2.10.0
python-docx>=0.8.11
youtube-transcript-api>=0.6.1
duckduckgo-search>=2.6.0
scikit-learn>=1.0.2
numpy>=1.22.0
openai>=0.27.0
requests>=2.28.0
# Optional (install if you need better embeddings / OCR / faiss):
# sentence-transformers
# faiss-cpu
# pytesseract
# unstructured
